AddContract finally works!
Massive overhaul to fix the botched addContract. This version fixed the ugly implementation, and it even adds Extras to the ContractHasExtra table. Furthermore, this update also changed the delete methods, so extras are deleted along with their contract. Lastly, I've also added functionality to prevent the user from inputting a an start Date that's later than the end date, when choosing the rental period